refactor(ChatsListItem): add useChatsListItem
